requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0
django-braces==1.15.0
django-ckeditor==6.7.0
django-js-asset==2.2.0
django-model-utils==4.3.1
django-render-partial==0.4
django-star-ratings==0.9.2
djangorestframework==3.14.0
drf-spectacular==0.27.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
Pillow==10.1.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.16.2
sorl-thumbnail==12.10.0
sqlparse==0.4.4
swapper==1.3.0
tzdata==2023.3
uritemplate==4.1.1
